requirements.txt
python-telegram-bot == 5.3
python-dotenv == 0.6